restore home page and fragment API
