+++ src/pages/index.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <script type="module" src="../components/greeting.js"></script>
+    <script type="module" src="../components/card.js"></script>
 
     <!-- Greeting API Setup -->
     <script>
@@ -24,22 +24,16 @@
         let offset = 0 - page;
 
         globalThis.document.getElementById('load-products').addEventListener('click', async () => {
-          const data = await fetch(`/api/greeting?name=owen`).then(resp => resp.text());
-          console.log({ data });
-          // offset = offset += page;
-          // const html = await fetch(`/api/fragment?offset=${offset}&limit=10`).then(resp => resp.text());
-          // console.log({ html });
+          offset = offset += page;
+          const html = await fetch(`/api/fragment?offset=${offset}&limit=10`).then(resp => resp.text());
 
-          // document.getElementById('load-products-output').insertAdjacentHTML('beforeend', html);
+          document.getElementById('load-products-output').insertAdjacentHTML('beforeend', html);
         });
       });
     </script>
   </head>
 
   <body>
-
-    <app-greeting></app-greeting>
-    <app-greeting name="Greenwood"></app-greeting>
     <article>
       <h2>JSON API</h2>
       <p>This is an example of a Greenwood API Route returning JSON when fetched on-submit of the form to display a message on the page.  You can see it fire in the network tab as <i>/api/greeting</i></p>
