--- src/services/message.js
@@ -1,7 +1,0 @@
-function getMessage(name) {
-  return `Hello ${name}!`;
-}
-
-export {
-  getMessage
-}
